requirements.txt
openai
python-dotenv
selenium
requests
